removed png files and reader function, used csv instead
